fix(client): enable to render previous element in slot (experimental) (#696)
If we have `main.tsx`, we can place an error boundary there.
However with managed mode or nested routes, we may want to place an
error boundary in RSC.
This enables to use previous elements on client with some condition. It
only works with "static" layout.
This is experimental. We should add tests for this to keep the
capability, if possible.