[fix] File upload (#850)
I have a simple HTML form which submits a file to a server action via
multipart form upload. From the client side, the file looks proper, but
when the server action receives it, the file is much larger? Is there
some intermediate processing of this file or improper handling in the
library?

client sent:
```
File {
  lastModified: 1724514738000
  name: "bara.jpeg"
  size: 2094593
  type: "image/jpeg"
}
```

server received:
```
File {
  size: 3715003,
  type: 'image/jpeg',
  name: 'bara.jpeg',
  lastModified: 1724698582443
}
```

I went through the code, and the bug largely was parsing the body as a
string. Instead we have an array buffer so we can manipulate it more
directly. It's possible it may still work as a string, but I believe the
array buffer leaves more flexibility.

I've only tested directly in the application I've been building

---------

Co-authored-by: Daishi Kato <dai-shi@users.noreply.github.com>